🎨 Renames content/pages to content/pageMetadata
